# chore: optimize performance of `Assets` unit tests